% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNCREST.R
\name{HGNCREST}
\alias{HGNCREST}
\alias{hgnc_info}
\alias{searchableFields}
\alias{hgnc_fetch}
\alias{hgnc_search}
\title{Functions for querying the HGNC REST API}
\usage{
hgnc_info()

searchableFields(service = c("HGNC", "VGNC"))

hgnc_fetch(searchableField, value)

hgnc_search(searchableField = NULL, query)
}
\arguments{
\item{service}{\code{character(1)} Either "HGNC" or "VGNC".}

\item{searchableField}{\code{character(1)} A field as provided by
\code{searchableFields()}.}

\item{value}{\code{character(1)} A value to search for in the \code{searchableField}.}

\item{query}{\code{character()} A query string to search for in the HGNC database.
The query string can be a single string or a vector of strings. If the
query is a vector, the strings are concatenated with a "+". The query
string can contain logical operators such as "AND", "OR", and "NOT". The
query string can also contain field-value pairs separated by a colon. For
example, to search for genes with the symbol starting with "ZNF" and the
status "Approved", the query string would be c("ZNF*", "AND",
"status:Approved").}
}
\value{
hgnc_info: A list of general information about the HGNC database. It
includes \code{searchableFields} and \code{storedFields} metadata.

searchableFields: A character vector of searchable fields in the
HGNC or VGNC database.

hgnc_fetch: A \code{tibble} with information about the gene specified by
the \code{searchableField} and \code{value} arguments.

hgnc_search: A \code{data.frame} with information about the genes that
match the query.
}
\description{
The functions follow the HGNC REST API documentation at
\url{https://www.genenames.org/help/rest/}. There are three main operations
that can be performed with this package: fetching information about a
specific gene (\code{hgnc_fetch}), searching for genes based on a query
(\code{hgnc_search}), and fetching general information about the HGNC database
(\code{hgnc_info}).
}
\examples{
hgnc_info()
searchableFields()
hgnc_fetch("ena", "BC040926")
hgnc_search(query = "BRAF")
hgnc_search("symbol", "ZNF*") |>
    head()
hgnc_search("symbol", c("ZNF*", "AND", "status:Approved")) |>
    head()
hgnc_search("symbol", c("ZNF3", "OR", "ZNF12"))
hgnc_search("symbol", c("ZNF*", "NOT", "status:Approved")) |>
    head()
hgnc_search("symbol", c("ZNF*", "AND", 'status:Entry withdrawn')) |>
    head()
hgnc_search(
    "name",
    c("MAPK interacting", "AND", "locus_type:gene with protein product")
)
}
