% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGNCREST.R
\name{VGNCREST}
\alias{VGNCREST}
\alias{vgnc_info}
\alias{vgnc_fetch}
\alias{vgnc_search}
\title{Functions for querying the VGNC REST sister API}
\usage{
vgnc_info()

vgnc_fetch(searchableField, value)

vgnc_search(searchableField = NULL, query)
}
\arguments{
\item{searchableField}{\code{character(1)} A field as provided by
\code{searchableFields()}.}

\item{value}{\code{character(1)} A value to search for in the \code{searchableField}.}

\item{query}{\code{character()} A query string to search for in the HGNC database.
The query string can be a single string or a vector of strings. If the
query is a vector, the strings are concatenated with a "+". The query
string can contain logical operators such as "AND", "OR", and "NOT". The
query string can also contain field-value pairs separated by a colon. For
example, to search for genes with the symbol starting with "ZNF" and the
status "Approved", the query string would be c("ZNF*", "AND",
"status:Approved").}
}
\value{
vgnc_info: A list of general information about the VGNC database. It
includes \code{searchableFields} and \code{storedFields} metadata.

vgnc_fetch: A \code{tibble} with information about the gene specified by
the \code{searchableField} and \code{value} arguments.

vgnc_search: A \code{data.frame} with information about the genes that
match the \code{query} in the \code{searchableField}.
}
\description{
The functions follow the VGNC REST API documentation at
\url{https://www.genenames.org/help/rest/}. There are three main operations
that can be performed with this package: fetching information about a
specific gene (\code{vgnc_fetch}), searching for genes based on a query
(\code{vgnc_search}), and fetching general information about the VGNC database
(\code{vgnc_info}).
}
\examples{
vgnc_info()
vgnc_fetch("symbol", "ZNF3")
vgnc_search(query = "BRAF")
vgnc_search("symbol", "ZNF*") |>
    head()
vgnc_search("symbol", c("ZNF*", "AND", "status:Approved")) |>
    head()
}
